Refactored the meal request function and split functions into meal responses
