+++ index.html
@@ -54,6 +54,7 @@
 
       <h2>Foods</h2>
       <button id="new_food_button">Create New</button>
+      <input type="button" onclick="location.href='/foods.html';" value="Create New" />
       <h4>Add Selected To</h4>
       <div class="row">
         <div class="column-50">
